M3: move approval domain types onto repository layer
Per ADR-001, repositories are the canonical persistence layer. Move
ApprovalRequest + SubmitApprovalParams onto repository/types.ts (which
already housed ApprovalStatus/Action/Context) so the repository tree
no longer transitively depends on the deprecated approval-store.

Update sqlite/postgres/event-wrapper/stores-interfaces import paths.
No behaviour change.
