+++ packages/data-layer/src/repository/postgres/approval.ts
@@ -3,7 +3,7 @@ import { randomUUID } from 'crypto';
 import { toJsonColumn } from '../json-column.js';
 import { approvals } from '../../db/schema.js';
 import type { ApprovalScopeFilter, IApprovalRequestRepository } from '../interfaces.js';
-import type { ApprovalAction, ApprovalContext, ApprovalRequest, ApprovalStatus } from '../../stores/approval-store.js';
+import type { ApprovalAction, ApprovalContext, ApprovalRequest, ApprovalStatus } from '../types.js';
 
 type ApprovalRow = typeof approvals.$inferSelect;
